backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
azure-cosmos==4.5.1
azure-storage-blob==12.19.0
openai==1.12.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx==0.26.0
